+++ WhistleWind.AbnormalSigils/Abilities/Stat Icons/variableTime.cs
@@ -25,7 +25,7 @@ namespace WhistleWind.AbnormalSigils
         private void StatIcon_Time()
         {
             const string rulebookName = "Passing Time";
-            const string rulebookDescription = "The value represented with this sigil will be equal to the number of turns that have passed since this card was played.";
+            const string rulebookDescription = "The value represented with this sigil will be equal to the number of turns that have passed since this card was placed on the board.";
             Time.icon = AbilityHelper.CreateStatIcon<Time>(pluginGuid,
                 "sigilTime", rulebookName, rulebookDescription, true, false).Id;
         }