+++ WhistleWindLobotomyMod/Abilities/Special Abilities/Adoration.cs
@@ -1,6 +1,8 @@
 ﻿using DiskCardGame;
+using InscryptionAPI.Card;
 using System.Collections;
 using UnityEngine;
+using WhistleWind.AbnormalSigils;
 using WhistleWind.Core.Helpers;
 using WhistleWindLobotomyMod.Core.Helpers;
 
@@ -12,64 +14,69 @@ namespace WhistleWindLobotomyMod
         public SpecialTriggeredAbility SpecialAbility => specialAbility;
 
         public const string rName = "Adoration";
-        public const string rDesc = "At the start of the owner's turn, if Melting Love's Health is 1, absorb the Health of adjacent Slimes.";
+        public const string rDesc = "At the start of the owner's turn, if Melting Love has lost Health, absorb the Health of adjacent Slime cards until Melting Love is at max Health.";
 
         public override bool RespondsToUpkeep(bool playerUpkeep) => base.PlayableCard.OpponentCard != playerUpkeep;
 
         public override IEnumerator OnUpkeep(bool playerUpkeep)
         {
-            if (base.PlayableCard.Health > 1)
+            if (base.PlayableCard.Health >= base.PlayableCard.MaxHealth)
                 yield break;
 
-            if (Singleton<ViewManager>.Instance.CurrentView != View.Board)
-            {
-                Singleton<ViewManager>.Instance.SwitchToView(View.Board);
-                yield return new WaitForSeconds(0.15f);
-            }
+            yield return HelperMethods.ChangeCurrentView(View.Board);
+
             CardSlot leftSlot = Singleton<BoardManager>.Instance.GetAdjacent(base.PlayableCard.Slot, true);
             CardSlot rightSlot = Singleton<BoardManager>.Instance.GetAdjacent(base.PlayableCard.Slot, false);
 
-            // check if left card is valid target
-            bool leftValid = leftSlot != null && leftSlot.Card != null && leftSlot.Card.Info.name == "wstl_meltingLoveMinion";
-
-            // check if right card is valid target
-            bool rightValid = rightSlot != null && rightSlot.Card != null && rightSlot.Card.Info.name == "wstl_meltingLoveMinion";
+            bool leftValid = leftSlot?.Card != null && leftSlot.Card.HasTrait(AbnormalPlugin.LovingSlime) && leftSlot.Card.Info.name != base.PlayableCard.Info.name;
+            bool rightValid = rightSlot?.Card != null && rightSlot.Card.HasTrait(AbnormalPlugin.LovingSlime) && rightSlot.Card.Info.name != base.PlayableCard.Info.name;
 
             // break if none are valid
             if (!leftValid && !rightValid)
                 yield break;
 
-            if (leftValid)
+            // find out how much Health we can heal back
+            int missingHealth = base.PlayableCard.MaxHealth - base.PlayableCard.Health;
+            if (leftValid && missingHealth > 0)
             {
-                leftSlot.Card.Anim.LightNegationEffect();
+                // if we need to absorb more health than the left card has, only heal by the left card's current Health
+                int healAmount = (leftSlot.Card.Health - missingHealth) > 0 ? missingHealth : leftSlot.Card.Health;
+                missingHealth -= healAmount; // update our remaining health we need to heal back
+
+                leftSlot.Card.Anim.StrongNegationEffect();
+                yield return new WaitForSeconds(0.4f);
+                leftSlot.Card.Status.damageTaken += healAmount;
+                leftSlot.Card.UpdateStatsText();
+                if (leftSlot.Card.Health > 0)
+                    leftSlot.Card.Anim.PlayHitAnimation();
+                else
+                    yield return leftSlot.Card.Die(false);
+
+                base.Card.Anim.LightNegationEffect();
+                base.PlayableCard.HealDamage(healAmount);
                 yield return new WaitForSeconds(0.2f);
+                    
             }
-            if (rightValid)
+            // if rightValid and we still have missing health, absorb from the right card as well
+            if (rightValid && missingHealth > 0)
             {
-                rightSlot.Card.Anim.LightNegationEffect();
+                int healAmount = (rightSlot.Card.Health - missingHealth) > 0 ? missingHealth : rightSlot.Card.Health;
+                
+                rightSlot.Card.Anim.StrongNegationEffect();
+                yield return new WaitForSeconds(0.4f);
+                rightSlot.Card.Status.damageTaken += healAmount;
+                rightSlot.Card.UpdateStatsText();
+                if (leftSlot.Card.Health > 0)
+                    rightSlot.Card.Anim.PlayHitAnimation();
+                else
+                    yield return rightSlot.Card.Die(false);
+
+                base.Card.Anim.LightNegationEffect();
+                base.PlayableCard.HealDamage(healAmount);
                 yield return new WaitForSeconds(0.2f);
             }
 
             yield return DialogueHelper.PlayDialogueEvent("MeltingLoveAbsorb");
-
-            if (leftValid)
-            {
-                int leftHealth = leftSlot.Card.Health;
-                base.PlayableCard.Anim.StrongNegationEffect();
-                yield return leftSlot.Card.Die(false, base.PlayableCard);
-                base.PlayableCard.HealDamage(leftHealth);
-                yield return new WaitForSeconds(0.4f);
-            }
-            if (rightValid && base.PlayableCard.Health < base.PlayableCard.MaxHealth)
-            {
-                int rightHealth = rightSlot.Card.Health;
-                base.PlayableCard.Anim.StrongNegationEffect();
-                yield return rightSlot.Card.Die(false, base.PlayableCard);
-                base.PlayableCard.HealDamage(rightHealth);
-                yield return new WaitForSeconds(0.4f);
-            }
-            Singleton<ViewManager>.Instance.SwitchToView(View.Default);
-            yield return new WaitForSeconds(0.15f);
         }
     }
     public class RulebookEntryAdoration : AbilityBehaviour